moved RELATIONSHIP_DIR_CHECK_REGEX to rel file
