+++ wiggle_query_language/clauses/parsing_helpers/parse_statement_checks.py
@@ -13,6 +13,9 @@ from wiggle_query_language.clauses.regexes.make.make_patterns import (
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
     MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX,
     PARAM_LIST_VALUE_REGEX,
+)
+
+from wiggle_query_language.clauses.regexes.patterns.relationships import (
     RELATIONSHIP_DIR_CHECK_REGEX,
 )
