+++ wiggle_query_language/clauses/regexes/patterns/patterns_helpers.py
