+++ wiggle_shell/core/query/query.py
@@ -8,6 +8,7 @@ from wiggle_query_language.clauses.make import (
     make,
     parse_make_statement_from_query_string,
 )
+from wiggle_query_language.clauses.find import parse_find_statement_from_query_string
 
 
 #  todo docstrings
@@ -21,7 +22,7 @@ def parse_query_string(query_string: str) -> ParsedQuery:
     """
 
     make_parsed = parse_make_statement_from_query_string(query_string)
-    find_parsed = None
+    find_parsed = parse_find_statement_from_query_string(query_string)
     criteria_parsed = None
     report_parsed = None
